Check/use cached Firebase CLI credential before Application Default Credentials.
On Cloudtop, the metadata server returns a default service account, so
authentication with Application Default Credentials is attempted. But that
account doesn't have any scopes, so authentication fails. The cached
Firebase CLI credentials are not used in that case.

By changing the order, authenticating with the cached Firebase CLI
credentials works on Cloudtops.