Fix upload performance.
The previous code passed the binary to be uploaded as a String object
(instead of an IO object) as the request body. This significantly
slowed upload performance since the SSL socket sends 16K at a time and
resizes the String buffer with each iteration. This caused CPU to
skyrocket since it requires allocating a new buffer and copying
existing data into that string. The garbage collector had to work hard
to keep up.

To eliminate this unnecessary String resizing, we now simply pass the
File object (which is an IO object) as the body, This enables
OpenSSL::Buffering to read and write a full 16K buffer. No resizing
(allocating, copying, garbage collecting) necessary.